requirements.txt
openai
wandb
tika
pysbd
pyspellchecker
torch
numpy
pandas
spacy